WIP: refactor group.ts
